# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A payment processing service that handles checkout sessions, subscription management, and webhooks",
        "title": "Payment Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/checkout": {
            "post": {
                "description": "Creates a new checkout session for a customer to purchase products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create checkout session",
                "parameters": [
                    {
                        "description": "Checkout request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetCheckoutUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout URL",
                        "schema": {
                            "$ref": "#/definitions/dto.UrlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/subscriptions": {
            "post": {
                "description": "Retrieves a link for customers to manage their subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get subscription management link",
                "parameters": [
                    {
                        "description": "Customer ID request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetSubscriptionManagementLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Management URL",
                        "schema": {
                            "$ref": "#/definitions/dto.UrlResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/webhook": {
            "post": {
                "description": "Processes webhook events from Dodo Payments and forwards data to configured service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Handle webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook signature for verification",
                        "name": "webhook-signature",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique webhook identifier",
                        "name": "webhook-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Webhook payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DodoWebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid signature",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.DodoWebhookPayload": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "customer": {
                            "$ref": "#/definitions/models.Customer"
                        },
                        "metadata": {
                            "type": "object",
                            "additionalProperties": {}
                        },
                        "product_id": {
                            "type": "string"
                        },
                        "status": {
                            "type": "string"
                        },
                        "subscription_id": {
                            "type": "string"
                        }
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.GetCheckoutUrlRequest": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/models.CustomerInput"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "product_cart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                }
            }
        },
        "dto.GetSubscriptionManagementLinkRequest": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                }
            }
        },
        "dto.UrlResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CustomerInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        }
    }
}